Adds script to chown admin:admin Encoder, Decoder directories
This is a workaround to the permission problem I have when using
robocopy with cygwin.

When I do this from a cygwin shell:

robocopy . //10.0.2.10/Production/Streambox/Beta/Decoder/3.140.01 \
	decoderHD-3.140.02-web223-ccf3223.exe

The the decoderHD-3.140.02-web223-ccf3223.exe file becomes unaccessible
to anyone but me.  chown admin:admin
decoderHD-3.140.02-web223-ccf3223.exe fixes this.

So this alias is a workaround to the robocopy-with-cygwin permission
problem by recursively re-setting ownership to admin:admin on the
readynas.